% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/education.r
\name{buildEducation}
\alias{buildEducation}
\title{buildEducation}
\usage{
buildEducation(info)
}
\arguments{
\item{info}{A list of lists for school information.  Each inner list must be of the form c('school', 'dates', 'degree', 'gpa') as in \code{\link{schoolText}}.}
}
\description{
Builds an education section for all schools
}
\details{
Builds LaTeX formated code for all schools
}
\examples{
eduction <- list(
list(school='Columbia University',
dates=c(2007, 2009), degree="Master of Arts in Statistics", gpa=3.926),
list(school='Muhlenberg College',
dates=c(2000, 2004),
degree="Bachelor of Science in Mathematics, Business Administration Minor", gpa=3.748))
buildEducation(education)
}
\author{
Jared P. Lander
}

