% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateDatabase.r
\name{createDatabase}
\alias{createDatabase}
\title{createDatabase}
\usage{
createDatabase(location = getwd(), name = "Resume")
}
\arguments{
\item{location}{Directory in which to create the database}

\item{name}{Name to use for the database}
}
\description{
Creates a database to hold resume information
}
\details{
Creates a SQLITE database with the necessary columns for making a resume
}
\author{
Jared P. Lander
}

