% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateJobFile.r
\name{createJobFile}
\alias{createJobFile}
\title{createJobFile}
\usage{
createJobFile(filename = "Resume.csv", sep = ",")
}
\arguments{
\item{filename}{Name of file in which to create the csv}

\item{sep}{Separator to use, ; is suggested}
}
\value{
An empty data.frame
}
\description{
Creates a CSV to hold information about jobs and research
}
\details{
This creates a data.frame and writes an empty file to disk.  This file should either be edited by hand or with a data.frame.
}
\examples{
\dontrun{
createJobFile() 
}

}
\author{
Jared P. Lander
}
