% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grades.R
\docType{data}
\name{grades}
\alias{grades}
\title{Grades obtained from the course of General Biostatistics.}
\format{This data frame contains 47 rows and 2 columns:
\itemize{
  \item{\code{score}} grades by individual.
  \item{\code{year}} grouped in year.
}}
\usage{
data(grades)
}
\description{
Grades obtained from the course of General Biostatistics.
}
\references{
Grades obtained from the course of General Biostatistics,courses from Universidad Nacional, Costa Rica.
}
\author{
Oscar Ramirez Alan \email{osoramirez@gmail.com}
}
\keyword{datasets}
\keyword{grades}
