% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs-data.r
\docType{data}
\name{jobs}
\alias{jobs}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 27 rows and 10 variables:
\describe{
  \item{JobName}{The internal name given to ID the job}
  \item{Company}{Name of company}
  \item{Location}{Job Location}
  \item{Title}{Position Title}
  \item{Start}{Start date of job}
  \item{End}{End date of job}
  \item{Bullet}{Bullet points for jobs}
  \item{BulletName}{Name or ID for bullets}
  \item{Type}{Type of job, either a job or research}
  \item{Description}{Short blurb about the job}
}}
\source{
Manufactured data
}
\usage{
jobs
}
\description{
A dataset containing the listings for a resume
}
\keyword{datasets}

