% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbox.R
\name{plotbox}
\alias{plotbox}
\title{plotbox}
\usage{
plotbox(x)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}
}
\value{
a plotbox (a elegant boxplot)
}
\description{
get a boxplot, showing the mean in plot
}
\examples{
x<-rnorm(25,2,3)
plotbox(x)
}
\keyword{plotbox}
