% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumendf.R
\name{resumendf}
\alias{resumendf}
\title{A Summary Statitical for Data.frame}
\usage{
resumendf(formula, data)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}
}
\value{
a table summary
}
\description{
The table give a summary statistics
}
\examples{
resumendf(conc~Treatment, data = CO2)

}
\keyword{resumendf}
