% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumeR-package.R
\docType{package}
\name{resumeR-package}
\alias{resumeR}
\alias{resumeR-package}
\title{resumeR: The Most Complete Summary Statistic Ever!}
\description{
Estimates complete summary statistics from 
  vector data.  Includes statistics for measures of central 
  tendency and spread.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/osoramirez/resumeR}
  \item Report bugs at \url{https://github.com/osoramirez/resumeR/issues}
}

}
\author{
\strong{Maintainer}: Oscar Ramirez Alan \email{osoramirez@gmail.com}

}
\keyword{internal}
