% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatFunction.r
\name{resumer}
\alias{resumer}
\title{resumer}
\usage{
resumer(fig_width = 4, fig_height = 2.5, fig_crop = TRUE, dev = "pdf",
  highlight = "default", keep_tex = FALSE, latex_engine = "pdflatex",
  includes = NULL, md_extensions = NULL, pandoc_args = NULL,
  template = "default")
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility (if available) to pdf figures}

\item{dev}{Graphics device to use for figure output (defaults to pdf)}

\item{highlight}{Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are "pdflatex", "lualatex", and "xelatex".}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}}
for additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{template}{Pandoc template to use for rendering. Pass "default" to use the resumer package default template; pass \code{NULL} to use pandoc's built-in template; pass a path to use a custom template that you've created.}
}
\description{
Convert to a resume
}

