% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/education.r
\name{schoolText}
\alias{schoolText}
\title{schoolText}
\usage{
schoolText(schoolInfo)
}
\arguments{
\item{schoolInfo}{A named list of information about the school.  Must contain c('school', 'dates', 'degree', 'gpa')}
}
\value{
LaTeX formatted text for the school
}
\description{
Builds LaTeX code for info on a school
}
\details{
Builds LaTeX code for info on a school
}
\examples{
schoolText(list(school='Columbia University',
dates=c(2007, 2009), degree="Master of Arts in Statistics", gpa=3.926))
}
\author{
Jared P. Lander
}

