% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumeR-package.R
\docType{package}
\name{resumeR-package}
\alias{resumeR}
\alias{resumeR-package}
\title{resumeR: The most complete summary statistic ever!}
\description{
This package is very useful, to get a complety summary statistic
We can get a plots, and see what it happen with your data. Only work for a vector data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/osoramirez/resumeR}
  \item Report bugs at \url{https://github.com/osoramirez/resumeR/issues}
}

}
\author{
\strong{Maintainer}: Oscar Ramirez Alan \email{osoramirez@gmail.com}

}
\keyword{internal}
